Move demo classed to a separate folder
