+++ src/main/java/demo/entity/Note.java
@@ -1,0 +1,23 @@
+package demo.entity;
+
+import lombok.Data;
+import org.svydovets.annotation.Column;
+import org.svydovets.annotation.Entity;
+import org.svydovets.annotation.Id;
+import org.svydovets.annotation.Table;
+
+@Entity
+@Data
+@Table(name = "notes")
+public class Note {
+
+    @Id
+    private Long id;
+
+    private String title;
+
+    private String body;
+
+    @Column(name = "user_id")
+    private Long userId;
+}
