+++ src/main/java/demo/entity/User.java
@@ -1,0 +1,24 @@
+package demo.entity;
+
+import lombok.Data;
+import org.svydovets.annotation.Column;
+import org.svydovets.annotation.Entity;
+import org.svydovets.annotation.Id;
+import org.svydovets.annotation.Table;
+
+@Entity
+@Data
+@Table(name = "users")
+public class User {
+
+    @Id
+    private Long id;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    private int age;
+}
